@@ -1,20 +1,19 @@
-import * as React from 'react'
-import Button from '@material-ui/core/Button'
+import React from 'react'
+import Button from '@mui/material/Button'
 
 import './Card.less'
 
 export default function Card(props: {
 	children: string
 	selected?: boolean
-	disabled?: boolean
-	onClick?: (score: string) => void
+	onClick?: () => void
 }) {
 	return (
 		<Button
 			className='card'
 			color='primary'
 			variant={props.selected ? 'contained' : 'outlined'}
-			onClick={props.disabled ? undefined : () => { props.onClick(props.children) }}
+			onClick={props.onClick}
 		>
 			{props.children}
 		</Button>
