@@ -1,25 +1,17 @@
-import * as _ from 'lodash'
+import { words, get } from 'lodash-es'
 
-const emailToAcronymMap = {
+const predefinedEmailToAcronymMap: Record<string, string | undefined> = {
 	'anantachai.s@taskworld.com': 'MT',
-	'chaiyapon.o@taskworld.com': 'NS',
-	'jacek.w@taskworld.com': 'JC',
-	'kunanan.t@taskworld.com': 'OL',
-	'sawit.r@taskworld.com': 'GR',
-	'shenaz.s@taskworld.com': 'SN',
-	'totsawat.m@taskworld.com': 'MX',
-	'manish.k@taskworld.com': 'MN',
-	'marek@taskworld.com': 'MR',
-	'phakamas.j@taskworld.com': 'JN',
 }
 
-export function getAcronym(email: string) {
-	if (emailToAcronymMap[email]) {
-		return emailToAcronymMap[email]
+export function getAcronym(email: string): string {
+	const predefinedAcronym = predefinedEmailToAcronymMap[email]
+	if (predefinedAcronym) {
+		return predefinedAcronym
 	}
 
 	const emailName = email.split('@')[0]
-	const [firstWord, secondWord] = _.words(emailName)
+	const [firstWord, secondWord] = words(emailName)
 	if (secondWord) {
 		return (firstWord.charAt(0) + secondWord.charAt(0)).toUpperCase()
 	}
@@ -27,7 +19,7 @@ export function getAcronym(email: string) {
 	const nonVowel = /[^aeiou_\-\.]/ig
 	const firstChar = /^[aeiou]/i.test(firstWord)
 		? firstWord.charAt(0)
-		: _.get(firstWord.match(nonVowel), '0', firstWord.charAt(0))
-	const secondChar = _.get(firstWord.match(nonVowel), '1', firstWord.charAt(1) || '')
+		: get(firstWord.match(nonVowel), '0', firstWord.charAt(0))
+	const secondChar = get(firstWord.match(nonVowel), '1', firstWord.charAt(1) || '')
 	return (firstChar + secondChar).toUpperCase()
 }
